Title: 각 마을로부터의 최소거리, Time: 892ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240530/각 마을로부터의 최소거리/minimum-distance-from-each-village.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main() {
     int n;
     cin >>n;
-    int x[n],a[n];
+    long x[n],a[n];
     for(int i=0;i<n;i++)
         cin>>x[i]>>a[i];
     int answer,tmp=1e10;
